refactor: share launcher package options
