+++ crates/launcher/src/lib.rs
@@ -9,6 +9,7 @@ mod package;
 #[cfg(not(any(target_os = "macos", test)))]
 #[path = "package_unsupported.rs"]
 mod package;
+mod package_options;
 mod supervisor;
 
 pub use error::{LauncherError, PackageError};
@@ -16,5 +17,6 @@ pub use layout::{
     BundleLayout, LAUNCHER_BUNDLE_IDENTIFIER, LAUNCHER_EXECUTABLE_NAME, OUTER_BUNDLE_NAME,
     WORKER_BUNDLE_IDENTIFIER, WORKER_BUNDLE_NAME, WORKER_EXECUTABLE_NAME,
 };
-pub use package::{PackageOptions, build_bundle};
+pub use package::build_bundle;
+pub use package_options::PackageOptions;
 pub use supervisor::{LauncherExit, launch_embedded_worker};
